figuring out how to test commonjs imports
