+++ gulp/tasks/copy-js.js
@@ -2,12 +2,10 @@
   'use strict';
 
   var gulp = require('gulp'),
-      $ = require('gulp-load-plugins')(),
       config = require('../config');
 
   gulp.task('copy-js', function () {
-    return gulp.src([config.app + '/backbone/**/*.js'])
-      .pipe(gulp.dest(config.dist + '/backbone'));
+    return gulp.src([config.app + '/**/*.js'])
+      .pipe(gulp.dest(config.dist));
   });
-
 })();
