+++ gulp/tasks/test.js
@@ -7,7 +7,7 @@
       SpecReporter = require('jasmine-spec-reporter'),
       argv = require('minimist')(process.argv.slice(3));
 
-  gulp.task('test', ['compressed-scripts', 'move-test', 'templates', 'test-vendor'], function() {
+  gulp.task('test', ['copy-js', 'move-test', 'templates', 'test-vendor'], function() {
     var testFile = argv.file || '*';
     return gulp.src([config.test + '/**/' + testFile + '.js'])
       .pipe($.jasmine({
