+++ package.json
@@ -9,6 +9,8 @@
     "backbone": "1.1.2",
     "backbone-nested": "2.0.1",
     "backbone.stickit": "0.8.0",
+    "browserify": "^9.0.3",
+    "commonjs-require": "^1.4.6",
     "gulp": "^3.8.8",
     "gulp-browserify": "^0.5.0",
     "gulp-clean": "^0.3.1",
@@ -17,19 +19,23 @@
     "gulp-jasmine": "2.0.0",
     "gulp-load-plugins": "^0.7.0",
     "gulp-plumber": "^0.6.6",
+    "gulp-rename": "^1.2.0",
     "gulp-uglify": "^1.0.1",
     "gulp-util": "^3.0.1",
     "gulp-wrap": "0.9.0",
+    "gulp-wrap-commonjs": "^0.1.13",
     "gulp-yuidoc": "^0.1.2",
     "handlebars": "3.0.0",
     "jasmine-reporters": "2.0.2",
     "jasmine-spec-reporter": "2.1.0",
     "jquery": "2.1.3",
+    "jquery-browserify": "^1.8.1",
     "jquery-mockjax": "1.6.1",
     "jsdom": "3.0.2",
     "minimist": "1.1.0",
     "promise": "6.1.0",
     "require-dir": "^0.1.0",
+    "requireify": "^0.2.1",
     "underscore": "1.8.2"
   },
   "engines": {
