+++ src/main/javascript/backbone/TorsoEvents.js
@@ -1,0 +1,2 @@
+
+    Torso.Events = _.extend({}, Backbone.Events);
