+++ src/main/javascript/backbone/collections/TorsoCollection.js
@@ -1,0 +1,1 @@
+Torso.Collection = Backbone.Collection.extend({});
