+++ src/main/javascript/backbone/mixins/TorsoCollectionLoadingMixin.js
