+++ src/main/javascript/backbone/mixins/TorsoCollectionRegistrationMixin.js
