+++ src/main/javascript/backbone/mixins/TorsoPollingMixin.js
