+++ src/main/javascript/backbone/mixins/TorsoViewHierarchyMixin.js
