+++ src/main/javascript/backbone/models/TorsoModel.js
@@ -1,0 +1,1 @@
+Torso.Model = Backbone.Model.extend({});
