+++ src/main/javascript/backbone/views/TorsoListView.js
