+++ src/main/javascript/backbone/views/TorsoView.js
@@ -1,0 +1,1 @@
+Backbone.View.extend({});
