+++ src/main/javascript/utils/torsoGuidManager.js
