+++ src/main/javascript/utils/torsoHandlebarsUtils.js
