+++ src/main/javascript/utils/torsoStickitUtils.js
