+++ src/main/javascript/utils/torsoTemplateRenderer.js
