+++ src/test/source/setupCommonJs.js
@@ -1,0 +1,4 @@
+window.$ = window.jQuery = require('jquery');
+window._ = require('underscore');
+window.Handlebars = require('handlebars');
+require('../../main/javascript/backbone/models/TorsoFormModel');
